+++ stsiapanau_bot/src/main/java/rent/RentHolder.java
@@ -27,28 +27,28 @@ public class RentHolder {
 	private double hotWaterRate;
 	private double outFallRate;
 	private String monthRent;
-	private long countT1;
-	private long usedT1;
+	private double countT1;
+	private double usedT1;
 	private double priceT1;
-	private long countT2;
-	private long usedT2;
+	private double countT2;
+	private double usedT2;
 	private double priceT2;
-	private long countT3;
-	private long usedT3;
+	private double countT3;
+	private double usedT3;
 	private double priceT3;
-	private long countColdWater;
-	private long usedColdWater;
+	private double countColdWater;
+	private double usedColdWater;
 	private double priceColdWater;
-	private long countHotWater;
-	private long usedHotWater;
+	private double countHotWater;
+	private double usedHotWater;
 	private double priceHotWater;
-	private long countOutFall;
+	private double countOutFall;
 	private double priceOutFall;
-	private long lastT1Count;
-	private long lastT2Count;
-	private long lastT3Count;
-	private long lastColdWaterCount;
-	private long lastHotWaterCount;
+	private double lastT1Count;
+	private double lastT2Count;
+	private double lastT3Count;
+	private double lastColdWaterCount;
+	private double lastHotWaterCount;
 	private Double total;
 	private boolean setLight;
 	private boolean setWater;
@@ -206,7 +206,7 @@ public class RentHolder {
 		this.monthRent = monthRent;
 	}
 
-	public long getCountT1() {
+	public double getCountT1() {
 		return this.countT1;
 	}
 
@@ -214,7 +214,7 @@ public class RentHolder {
 		this.countT1 = countT1;
 	}
 
-	public long getCountT2() {
+	public double getCountT2() {
 		return this.countT2;
 	}
 
@@ -222,7 +222,7 @@ public class RentHolder {
 		this.countT2 = countT2;
 	}
 
-	public long getCountT3() {
+	public double getCountT3() {
 		return this.countT3;
 	}
 
@@ -230,7 +230,7 @@ public class RentHolder {
 		this.countT3 = countT3;
 	}
 
-	public long getCountColdWater() {
+	public double getCountColdWater() {
 		return this.countColdWater;
 	}
 
@@ -238,7 +238,7 @@ public class RentHolder {
 		this.countColdWater = countColdWater;
 	}
 
-	public long getCountHotWater() {
+	public double getCountHotWater() {
 		return this.countHotWater;
 	}
 
@@ -246,23 +246,23 @@ public class RentHolder {
 		this.countHotWater = countHotWater;
 	}
 
-	public long getLastT1Count() {
+	public double getLastT1Count() {
 		return lastT1Count;
 	}
 
-	public long getLastT2Count() {
+	public double getLastT2Count() {
 		return this.lastT2Count;
 	}
 
-	public long getLastT3Count() {
+	public double getLastT3Count() {
 		return this.lastT3Count;
 	}
 
-	public long getLastColdWaterCount() {
+	public double getLastColdWaterCount() {
 		return this.lastColdWaterCount;
 	}
 
-	public long getLastHotWaterCount() {
+	public double getLastHotWaterCount() {
 		return this.lastHotWaterCount;
 	}
 
@@ -274,7 +274,7 @@ public class RentHolder {
 		this.setWater = setWater;
 	}
 
-	public long getUsedT1() {
+	public double getUsedT1() {
 		return this.usedT1;
 	}
 
@@ -282,7 +282,7 @@ public class RentHolder {
 		return priceT1;
 	}
 
-	public long getUsedT2() {
+	public double getUsedT2() {
 		return this.usedT2;
 	}
 
@@ -290,7 +290,7 @@ public class RentHolder {
 		return this.priceT2;
 	}
 
-	public long getUsedT3() {
+	public double getUsedT3() {
 		return this.usedT3;
 	}
 
@@ -298,7 +298,7 @@ public class RentHolder {
 		return this.priceT3;
 	}
 
-	public long getUsedColdWater() {
+	public double getUsedColdWater() {
 		return this.usedColdWater;
 	}
 
@@ -306,7 +306,7 @@ public class RentHolder {
 		return this.priceColdWater;
 	}
 
-	public long getUsedHotWater() {
+	public double getUsedHotWater() {
 		return this.usedHotWater;
 	}
 
@@ -314,7 +314,7 @@ public class RentHolder {
 		return this.priceHotWater;
 	}
 
-	public long getCountOutFall() {
+	public double getCountOutFall() {
 		return this.countOutFall;
 	}
